back_end/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
dotenv==0.9.9
idna==3.10
numpy==2.2.4
oauthlib==3.2.2
pillow==11.1.0
psycopg==3.2.6
psycopg-binary==3.2.6
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.13.0
urllib3==2.3.0
